# travaia-ai/backend/api-gateway/openapi.yaml
swagger: "2.0"
info:
  title: Travaia API Gateway
  description: API Gateway for Travaia microservices
  version: "1.0.0"
host: travaia-managed-cggfdfie.uc.gateway.dev
basePath: /
schemes:
  - https
produces:
  - application/json
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "key"
    in: "query"
  firebase:
    authorizationUrl: ""
    flow: "implicit"
    type: "oauth2"
    x-google-issuer: "https://securetoken.google.com/travaia-e1310"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
    x-google-audiences: "travaia-e1310"

# CORS configuration for cross-origin requests
...
paths:
...
  /profile:
    get:
      summary: Get user profile
      operationId: profile_get
      security:
        - firebase: []
      x-google-backend:
        address: https://travaia-user-auth-service-976191766214.us-central1.run.app/profile
      responses:
        '200':
          description: Success
    put:
      summary: Update user profile
      operationId: profile_put
      security:
        - firebase: []
      x-google-backend:
        address: https://travaia-user-auth-service-976191766214.us-central1.run.app/profile
      responses:
        '200':
          description: Success
    delete:
      summary: Delete user account
      operationId: profile_delete
      security:
        - firebase: []
      x-google-backend:
        address: https://travaia-user-auth-service-976191766214.us-central1.run.app/profile
      responses:
        '200':
          description: Success
  /profile/settings:
    get:
      summary: Get user settings
      operationId: profile_settings_get
      security:
        - firebase: []
      x-google-backend:
        address: https://travaia-user-auth-service-976191766214.us-central1.run.app/profile/settings
      responses:
        '200':
          description: Success
    put:
      summary: Update user settings
      operationId: profile_settings_put
      security:
        - firebase: []
      x-google-backend:
        address: https://travaia-user-auth-service-976191766214.us-central1.run.app/profile/settings
      responses:
        '200':
          description: Success
...
  /gamification/progress:
    get:
      summary: Get user progress
      operationId: gamification_progress_get
      security:
        - firebase: []
      x-google-backend:
        address: https://travaia-user-auth-service-976191766214.us-central1.run.app/gamification/progress
      responses:
        '200':
          description: Success
...
  /api/users/profile:
    put:
      summary: Update user profile data
      operationId: api_users_profile_put
      security:
        - firebase: []
      x-google-backend:
        address: https://travaia-user-auth-service-976191766214.us-central1.run.app/api/users/profile
      responses:
        '200':
          description: Success
  /api/users/progress:
    put:
      summary: Update user progress data
      operationId: api_users_progress_put
      security:
        - firebase: []
      x-google-backend:
        address: https://travaia-user-auth-service-976191766214.us-central1.run.app/api/users/progress
      responses:
        '200':
          description: Success
  /api/users/settings:
    put:
      summary: Update user settings data
      operationId: api_users_settings_put
      security:
        - firebase: []
      x-google-backend:
        address: https://travaia-user-auth-service-976191766214.us-central1.run.app/api/users/settings
      responses:
        '200':
          description: Success
...